fix(dpp): make document type index update validation name-order independent
IndexLevel::validate_update compared index trees by level_identifier
equality, but identifiers are assigned by an incrementing counter that
walks the document type's indices — a BTreeMap keyed by index NAME. An
index change whose name sorted before existing indexes renumbered every
level and was rejected with an opaque "Invalid path", while the
semantically identical change under a late-sorting name passed the tree
comparison and instead hard-errored later in the JSON-schema
compatibility check (which has no rule for the `indices` keyword),
surfacing as an InternalError tx result. Which consensus outcome a
contract owner got therefore depended purely on the index name string.

At protocol v14, validate_update v1 replaces the tree comparison with a
name-keyed comparison of the parsed index definitions: any added,
removed or modified index is rejected with a deterministic
DataContractInvalidIndexDefinitionUpdateError naming the offending
index, independent of name sort order. The accepted set is unchanged —
under v0 no index modification could ever pass the full pipeline, so
this only makes the rejection deterministic and correctly labeled.

Gated per the new-version-module convention: new
DPP_VALIDATION_VERSIONS_V5 (document_type.validate_update: 1) wired
into v14 only; the v0 flow is byte-for-byte frozen for protocol
versions <= 13 and pinned by a regression test. Generation-independent
helpers (validate_config, byte-array encoding stability,
validate_schema) move verbatim to a shared common module, following the
try_from_schema extraction precedent (#4276).
