+++ packages/rs-platform-version/src/version/dpp_versions/dpp_validation_versions/mod.rs
@@ -4,6 +4,7 @@ pub mod v1;
 pub mod v2;
 pub mod v3;
 pub mod v4;
+pub mod v5;
 
 #[derive(Clone, Debug, Default)]
 pub struct DPPValidationVersions {
